refactor(dimensions): streamline filters and dimension management flow
